# swagger/swaggerDoc.yaml
openapi: 3.0.0
info:
  title: "Your API Docs"
  description: "Your API Description"
  version: "1.0.0"
basePath: /api/v1
servers:
  - url: https://graduation-project-nrnm.onrender.com/
    description: Production server (uses live data)

  - url: 127.0.0.1:30
    description: local server (uses test data)
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /users:
    get:
      summary: Get all users
      tags:
        - Users
      responses:
        200:
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
      security:
        - bearerAuth: []
...
  /shops:
    get:
      summary: "Get all shops"
      tags:
        - Shops
      description: "Returns a list of all shops"
      responses:
        200:
          description: "Success"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Shop"
    post:
      summary: "Create a new shop"
      tags:
        - Shops
      description: "Creates a new shop"
      parameters:
        - name: "Shop"
          in: "body"
          description: "Shop object"
          required: true
          schema:
            $ref: "#/definitions/Shop"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/Shop"
...
definitions:
...
  Shop:
      type: object
      properties:
        id:
          type: integer
          description: The unique identifier for the shop
        name:
          type: string
          description: The name of the shop
        location:
          type: string
          description: The location of the shop
        openAt:
          type: integer
          description: The opening hour of the shop
        closeAt:
          type: integer
          description: The closing hour of the shop
        phone:
          type: string
          description: The phone number of the shop
        imageUrl:
          type: string
          description: The URL of the image for the shop
        shopType:
          type: string
          description: The type of the shop
          enum: [shop, restaurant, entreatment]
      required:
        - name
        - location
        - openAt
        - closeAt
        - phone
        - shopType
components:
  schemas:
    User:
      type: object
      required:
        - name
        - phone
        - email
        - password
        - passwordConfirm
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the user
        name:
          type: string
          description: User's full name
        phone:
          type: integer
          format: int64
          description: User's phone number
        email:
          type: string
          description: User's email address
        password:
          type: string
          description: User's password
          minLength: 8
          maxLength: 100
        passwordConfirm:
          type: string
          description: Password confirmation
          minLength: 8
          maxLength: 100
        passwordChangedAt:
          type: string
          format: date-time
          description: Timestamp of the last password change
        passwordResetCode:
          type: string
          description: Code used for resetting password
        passwordResetExpire:
          type: string
          format: date-time
          description: Timestamp when the password reset code expires
        imageURL:
          type: string
          description: URL of the user's profile picture
...